+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning][semver].
 ### Fixed
 
 - Fix nested jsonrpc message deserialization ([#140])
+- Fix server handling of client errors ([#141])
 
 ## [0.9.1] - 09/29/2020